refactor: reorganize claim institution files into feature directory
- Move server action from lib/actions/claim-request.ts to lib/features/claim-institution/actions.ts
- Move validation schema from lib/validations/claim-request.ts to lib/features/claim-institution/validations.ts
- Update import paths to use relative imports within feature directory
- Improve code organization by grouping related functionality together
