+++ components/ui/claim-modal.tsx
@@ -12,7 +12,7 @@ import {
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
-import { submitClaimRequest } from "@/lib/actions/claim-request";
+import { submitClaimRequest } from "@/lib/features/claim-institution/actions";
 import { useState } from "react";
 import { toast } from "sonner";
