+++ lib/features/claim-institution/actions.ts
@@ -3,10 +3,10 @@
 import { auth } from "@/auth";
 import { db } from "@/db";
 import { claimRequests, institutions } from "@/db/schema";
-import { claimRequestSchema } from "@/lib/validations/claim-request";
 import { and, eq } from "drizzle-orm";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
+import { claimRequestSchema } from "./validations";
 
 export async function submitClaimRequest(formData: FormData) {
 	try {
